requirements.txt
Flask==2.1.1
pyrebase4==4.5.0
requests==2.27.1
python-dotenv==0.20.0
Werkzeug==2.0.3
requests_toolbelt==0.9.1